pycraft-project-2013
One of my proudest projects: this is PyCraft. A 2D Top-Down rip-off of
Minecraft made by me in Python with PyGame. This was my final project
for Grade 11 Computer Science. Apparently making an infinite world
generator, game saving, getting player skins from the Minecraft skins
server, and moving a character in the direction of the mouse relative to
the centre of the screen among other things were very revolutionary for
our class since nothing of its sort was taught. The code is thoroughly
commented in pycraft.py